+++ kong/plugins/openid-connect/cache.lua
@@ -808,5 +808,5 @@ return {
   tokens         = tokens,
   token_exchange = token_exchange,
   userinfo       = userinfo,
-  version        = "0.1.7",
+  version        = "0.1.8",
 }