+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## x.x.x
+
+- Added the port parameter to be able to be set a new port for the alternate
+  upstream.
+
 ## 0.1.0
 
 - First commit for Kong plugin to handle Canary release of Kong APIs