+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## Unreleased
+
+### Fixed
+
+- Fix `require` statements pointing to CE ldap plugin
+- Fix usage of LuaJIT ffi; `ffi.load` was being used to access an external
+  symbol instead of `ffi.C.`
+
 ## 0.1.0
 
 ldap-auth-advanced was forked from Kong CE ldap-auth.